+++ apps/ekstra-omsorgsdager-andre-forelder-ikke-tilsyn/.storybook/preview.ts
@@ -1,5 +1,5 @@
 import { Preview } from '@storybook/react-vite';
-import '@navikt/ds-css';
+import '@navikt/ds-css/darkside';
 
 const preview: Preview = {
     globalTypes: {