+++ apps/ekstra-omsorgsdager-andre-forelder-ikke-tilsyn/src/app/søknad/steps/oppsummering/i18n/nb.ts
@@ -1,3 +1,4 @@
+/* eslint-disable max-len */
 export const oppsummeringMessages_nb = {
     'step.oppsummering.nextButtonLabel': 'Send inn søknaden',
     'step.oppsummering.info':
@@ -31,4 +32,15 @@ export const oppsummeringMessages_nb = {
     'step.oppsummering.sendSøknad': 'Send søknad',
 
     'validation.harBekreftetOpplysninger.notChecked': 'Du må bekrefte opplysningene.',
+
+    'innsendingFeilet.tittel': 'Oops, noe gikk galt.',
+    'innsendingFeilet.tekst.situasjonBeskrivelse':
+        'Feltet som beskriver situasjonen inneholder tegn som ikke er tillatt.',
+    'innsendingFeilet.tekst.generell.1': 'Søknaden din inneholder ugyldig informasjon.',
+    'innsendingFeilet.tekst.generell.2':
+        'Når du selv skriver inn tekst i et felt i søknaden, kan noen tegn være ugyldige ut fra informasjonen vi ber om. Dette skjer vanligvis hvis du kopierer og limer inn tekst fra andre steder. Du fikser dette ved å skrive inn teksten på ny direkte i feltet, uten å kopiere den fra et annet sted.',
+    'innsendingFeilet.tekst.generell.3':
+        'Noen ganger vil de ugyldige tegnene være tegn en ser på som vanlige, f.eks. emojier, «, » eller «§», mens andre ganger kan det være synlige som små firkanter inne i teksten.',
+    'innsendingFeilet.tekst.generell.4':
+        'Hvis du har sjekket dette, og fortsatt ikke kommer videre, ber vi deg kontakte oss på <Telefon>55 55 33 33</Telefon> for videre veiledning.',
 };