+++ apps/omsorgsdager-kalkulator/components/result/ResultatArea.tsx
@@ -21,14 +21,14 @@ const ResultatArea: React.FC<Props> = ({ resultView }: Props) => {
         () => null,
         () => (
             <ResultBox type={'WARNING'}>
-                <Box borderRadius="large" padding={'5'} background="surface-warning-moderate" className="mb-3">
+                <Box.New borderRadius="large" padding={'5'} background="warning-moderate" className="mb-3">
                     <Heading level="3" size="medium">
                         <FormattedMessage id={'resultat-area.green.2.1.a'} values={{ result: 0 }} />{' '}
                         <span className="font-normal">
                             <FormattedMessage id={'resultat-area.green.2.1.b'} values={{ year }} />
                         </span>
                     </Heading>
-                </Box>
+                </Box.New>
                 <BodyLong className="pt-4 pb-9 text-justify" size="small">
                     <FormattedMessage id={'resultat-area.orange.3'} />
                 </BodyLong>
@@ -47,7 +47,7 @@ const ResultatArea: React.FC<Props> = ({ resultView }: Props) => {
                 <>
                     <ResultBox type={'SUCCESS'}>
                         <VStack gap="4">
-                            <Box borderRadius="large" padding={'5'} background="surface-success-moderate">
+                            <Box.New borderRadius="large" padding={'5'} background="success-moderate">
                                 <Heading level="3" size="medium">
                                     <FormattedMessage
                                         id={'resultat-area.green.2.1.a'}
@@ -57,7 +57,7 @@ const ResultatArea: React.FC<Props> = ({ resultView }: Props) => {
                                         <FormattedMessage id={'resultat-area.green.2.1.b'} values={{ year }} />
                                     </span>
                                 </Heading>
-                            </Box>
+                            </Box.New>
 
                             <div className="subtle-card">
                                 <ExpansionCard