+++ apps/omsorgsdager-kalkulator/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @navikt/omsorgsdager-kalkulator
 
+## 3.12.38
+
+### Patch Changes
+
+- Pakkeoppdateringer
+
+## 3.12.37
+
+### Patch Changes
+
+- Pakkeoppdateringer
+
+## 3.12.36
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch + major uuid
+
+## 3.12.35
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 3.12.34
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
 ## 3.12.33
 
 ### Patch Changes