+++ packages/appstatus-sanity-config/CHANGELOG.md
@@ -1,5 +1,23 @@
 # appstatus-sanity-config
 
+## 1.2.29
+
+### Patch Changes
+
+- Pakkeoppdateringer
+
+## 1.2.28
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch + major uuid
+
+## 1.2.27
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
 ## 1.2.26
 
 ### Patch Changes