+++ packages/eslint-config-sif/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sif/eslint-config
 
+## 1.1.51
+
+### Patch Changes
+
+- Pakkeoppdateringer
+
+## 1.1.50
+
+### Patch Changes
+
+- Pakkeoppdateringer
+
+## 1.1.49
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch + major uuid
+
+## 1.1.48
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
 ## 1.1.47
 
 ### Patch Changes