+++ server-ungdomsytelse-veileder/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-server
 
+## 0.0.28
+
+### Patch Changes
+
+- Pakkeoppdateringer
+
+## 0.0.27
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
 ## 0.0.26
 
 ### Patch Changes